requirements.txt
numpy>=1.21.0
opencv-python>=4.7.0
pandas>=1.4.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
scikit-learn>=1.0.0
requests>=2.28.0
tqdm>=4.65.0
pytesseract>=0.3.10
fastapi
uvicorn